+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+ * v1.2.0 - Rewrites upload.html, adds footer to rendered markdown, in attempt to foster virality. Lots of improvements.
+ * v1.1.2 - Adds `&cors` URL parameter to add CORS headers to embed content from a getpost.
  * v1.1.1 - Bump small tweak for Python ergonomics and some packaging adjustments. Shoutout @technillogue
  * v1.1.0 - New beautiful CSS by @[iris-garden](https://github.com/iris-garden), more filetypes added to fix silly issues and bug found by @[ExAtrisUmbra](https://twitter.com/exastrisumbra).
  * v1.0.0 - Original Release - Feature Complete & Working -  Mar 6 13:34:35 2021